shop/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-js-asset==1.2.2
django-mptt==0.11.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
Pillow==8.1.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
